+++ libraries/ritual_arweave/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 - ##### The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - ##### This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2024-07-31
+- Upload/Download support for large files
+- Upload/Download support for generic blobs, as well as json dictionaries
+
 ## [0.1.0] - 2024-06-06
 
 ### Added